Add a read-only Redis replica in sjc
This replica will ensure that we have a copy of all Redis data on a
volume in the sjc region. If sea has an outage, we can edit a few config
files and make sjc the primary, at which point sea (when it returns)
will be the replica. We can then switch back whenever sea stabilizes.