Migrate Postgres to a 15.x flex cluster
The new pie-gd-postgres15 app is a 3-member Postgres 15 cluster using
Fly's new flex architecture.

Two members of the cluster are in the primary sea region and one is in
sjc, which provides us with redundancy in sea as well as the ability to
manually make sjc the primary region if it's necessary to fail out of
sea.

The migration was painless:

1.  Created the new pie-gd-postgres15 cluster

2.  Scaled pie-gd-mastodon down to 0 to stop writes

3.  Used `fly pg import` to import data from pie-gd-postgres to
    pie-gd-postgres15

4.  Detached pie-gd-postgres from pie-gd-mastodon

5.  Attached pie-gd-postgres15 to pie-gd-mastodon

6.  Scaled pie-gd-mastodon up to bring it back online