Start Elasticsearch on demand and shut it down when it's not needed
In theory this should significantly reduce the cost of running
Elasticsearch by automatically stopping the machine when it's not being
used and restarting it on demand when a request comes in.

The downside is there's a ~30 second delay while Elasticsearch starts
up, but then it'll stay running as long as requests are being made.

It remains to be seen whether this will cause any problems with
indexing. Let's find out!