Set Postgres volume auto extend threshold to 50%
With the threshold at 80%, it's possible for maintenance tasks like
backups or reindexing to rapidly use up the available free space before
volume extension is triggered.

Once disk utilization exceeds 90%, the database will be put into a
persistent read-only state that can only be resolved manually, even
after the volume is extended. This is not fun.

Extending the volume once it reaches 50% utilization will give us lots
of headroom and hopefully prevent this from happening.